ADD: scraper ozon is done
