+++ cmd/main-fu/main.go
@@ -1,10 +1,10 @@
 package main
 
-import "marketplace/internal/parcer"
+import "marketplace/internal/scraper"
 
 func main() {
-	scraber := parcer.NewScraber()
+	scraber := scraper.NewScraber()
 
 	scraber.GetOzon("https://www.ozon.ru/search/?text=носки&from_global=true")
-	scraber.GetOzon("https://www.ozon.ru/search/?text=bottel&from_global=true")
+	// scraber.GetOzon("https://www.ozon.ru/search/?text=bottel&from_global=true")
 }
