+++ internal/connect/connect.go
@@ -1,0 +1,15 @@
+package connect
+
+import (
+	"context"
+
+	"github.com/redis/go-redis/v9"
+)
+
+func ConnectToRedis() {
+	ctx := context.Background()
+
+	rbd := redis.NewClient(&redis.Options{
+		Addr: localhost,
+	})
+}
