--- src/reader/reader.go
@@ -1,5 +1,0 @@
-package reader
-
-type Reader interface {
-	Parse() (*Data, error)
-}
